Move DeleteKubeAPIServer function to kubeapiserver.go
