+++ pkg/operation/botanist/kubeapiserver_test.go
@@ -21,6 +21,8 @@ import (
 	"github.com/gardener/gardener/pkg/apis/core/v1beta1/helper"
 	seedmanagementv1alpha1 "github.com/gardener/gardener/pkg/apis/seedmanagement/v1alpha1"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
+	fakeclientmap "github.com/gardener/gardener/pkg/client/kubernetes/clientmap/fake"
+	"github.com/gardener/gardener/pkg/client/kubernetes/clientmap/keys"
 	"github.com/gardener/gardener/pkg/client/kubernetes/fake"
 	"github.com/gardener/gardener/pkg/features"
 	gardenletfeatures "github.com/gardener/gardener/pkg/gardenlet/features"
@@ -37,7 +39,9 @@ import (
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 	appsv1 "k8s.io/api/apps/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/component-base/featuregate"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -46,6 +50,8 @@ import (
 
 var _ = Describe("KubeAPIServer", func() {
 	var (
+		ctrl *gomock.Controller
+
 		c             client.Client
 		k8sSeedClient kubernetes.Interface
 		botanist      *Botanist
@@ -55,6 +61,8 @@ var _ = Describe("KubeAPIServer", func() {
 	)
 
 	BeforeEach(func() {
+		ctrl = gomock.NewController(GinkgoT())
+
 		c = fakeclient.NewClientBuilder().WithScheme(kubernetes.SeedScheme).Build()
 		k8sSeedClient = fake.NewClientSetBuilder().WithClient(c).Build()
 		botanist = &Botanist{
@@ -70,6 +78,10 @@ var _ = Describe("KubeAPIServer", func() {
 				},
 			},
 		}
+	})
+
+	AfterEach(func() {
+		ctrl.Finish()
 	})
 
 	Describe("#DefaultKubeAPIServer", func() {
@@ -289,18 +301,12 @@ var _ = Describe("KubeAPIServer", func() {
 
 	Describe("#DeployKubeAPIServer", func() {
 		var (
-			ctrl          *gomock.Controller
 			kubeAPIServer *mockkubeapiserver.MockInterface
 		)
 
 		BeforeEach(func() {
-			ctrl = gomock.NewController(GinkgoT())
 			kubeAPIServer = mockkubeapiserver.NewMockInterface(ctrl)
 			botanist.Shoot.Components.ControlPlane.KubeAPIServer = kubeAPIServer
-		})
-
-		AfterEach(func() {
-			ctrl.Finish()
 		})
 
 		DescribeTable("should correctly set the autoscaling replicas",
@@ -366,6 +372,34 @@ var _ = Describe("KubeAPIServer", func() {
 				int32(0),
 			),
 		)
+	})
+
+	Describe("#DeleteKubeAPIServer", func() {
+		It("should properly invalidate the client and delete the deployment", func() {
+			clientMap := fakeclientmap.NewClientMap().AddClient(keys.ForShoot(botanist.Shoot.Info), k8sSeedClient)
+			botanist.ClientMap = clientMap
+
+			shootClient, err := botanist.ClientMap.GetClient(ctx, keys.ForShoot(botanist.Shoot.Info))
+			Expect(err).NotTo(HaveOccurred())
+			Expect(shootClient).To(Equal(k8sSeedClient))
+
+			k8sShootClient := fake.NewClientSetBuilder().WithClient(c).Build()
+			botanist.K8sShootClient = k8sShootClient
+
+			deployment := &appsv1.Deployment{ObjectMeta: metav1.ObjectMeta{Name: "kube-apiserver", Namespace: shootNamespace}}
+			Expect(c.Create(ctx, deployment)).To(Succeed())
+			Expect(c.Get(ctx, client.ObjectKeyFromObject(deployment), deployment)).To(Succeed())
+
+			Expect(botanist.DeleteKubeAPIServer(ctx)).To(Succeed())
+
+			shootClient, err = clientMap.GetClient(ctx, keys.ForShoot(botanist.Shoot.Info))
+			Expect(err).To(MatchError("clientSet for key \"/\" not found"))
+			Expect(shootClient).To(BeNil())
+
+			Expect(botanist.K8sShootClient).To(BeNil())
+
+			Expect(c.Get(ctx, client.ObjectKeyFromObject(deployment), deployment)).To(MatchError(apierrors.NewNotFound(schema.GroupResource{Group: appsv1.SchemeGroupVersion.Group, Resource: "deployments"}, deployment.Name)))
+		})
 	})
 })
